move subquery to prepare
